+++ src/assets/styles/index.ts
@@ -1,0 +1,3 @@
+export * from './variables'
+export * from './Theme'
+export * from './OptionsGlobalStyles'