+++ src/store/actions/index.ts
@@ -1,2 +1,3 @@
 export * from './action-types'
 export * from './status'
+export * from './auth'