+++ webpack.config.common.babel.js
@@ -22,6 +22,18 @@ export default {
                 exclude: /node_modules/
             },
             {
+                test: /\.(woff(2)?|ttf|eot)(\?v=\d+\.\d+\.\d+)?$/,
+                use: [
+                    {
+                        loader: 'file-loader',
+                        options: {
+                            name: '[name].[ext]',
+                            outputPath: 'fonts/'
+                        }
+                    }
+                ]
+            },
+            {
                 test: /\.html$/,
                 loader: 'html-loader',
                 exclude: /node_modules/
@@ -47,10 +59,13 @@ export default {
         alias: {
             '@store': path.resolve(__dirname, 'src/store'),
             '@actions': path.resolve(__dirname, 'src/store/actions'),
-            '@reducers': path.resolve(__dirname, 'src/store/reducers/actions'),
+            '@reducers': path.resolve(__dirname, 'src/store/reducers'),
             '@components': path.resolve(__dirname, 'src/components'),
             '@containers': path.resolve(__dirname, 'src/containers'),
-            '@utils': path.resolve(__dirname, 'src/utils')
+            '@utils': path.resolve(__dirname, 'src/utils'),
+            '@api': path.resolve(__dirname, 'src/api'),
+            '@assets': path.resolve(__dirname, 'src/assets'),
+            '@styles': path.resolve(__dirname, 'src/assets/styles')
         }
     },
     plugins: [